last commit

+++ ft_putnbr_fd.c
@@ -6,27 +6,27 @@
 /*   By: rlarabi <rlarabi@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/10/14 21:12:33 by rlarabi           #+#    #+#             */
-/*   Updated: 2022/10/15 14:20:21 by rlarabi          ###   ########.fr       */
+/*   Updated: 2022/10/24 17:53:47 by rlarabi          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-void    ft_putnbr_fd(int n, int fd)
+void	ft_putnbr_fd(int n, int fd)
 {
-    long    l;
+	long	l;
 
-    l = n;
-    if (l < 0)
-    {
-        ft_putchar_fd('-',fd);
-        l *= -1;
-    }
-    if (l >= 10)
-    {
-        ft_putnbr_fd(l / 10, fd);
-        l %= 10;
-    }
-    if (l < 10)
-        ft_putchar_fd(l + '0', fd);
+	l = (long)n;
+	if (l < 0)
+	{
+		ft_putchar_fd('-', fd);
+		l *= -1;
+	}
+	if (l >= 10)
+	{
+		ft_putnbr_fd(l / 10, fd);
+		l %= 10;
+	}
+	if (l < 10)
+		ft_putchar_fd(l + '0', fd);
 }
