+++ ft_strlen.c
@@ -6,11 +6,13 @@
 /*   By: rlarabi <rlarabi@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/10/05 20:24:33 by rlarabi           #+#    #+#             */
-/*   Updated: 2022/10/20 12:16:08 by rlarabi          ###   ########.fr       */
+/*   Updated: 2022/10/24 21:50:41 by rlarabi          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-int	ft_strlen(const char *str)
+#include "libft.h"
+
+size_t	ft_strlen(const char *str)
 {
 	int	i;
